Split api tests
